added so much shit
Added: Growth Enchant
Added: ton of config settings
players can now change the look of the health bar when using Blessing of Knowledge via the config
you can also change the sounds that will play when an blacklisted item has been attempted to be applied
i have also added a blacklisted system
that can be toggled off
I have also added sounds for when there is already an existing enchant on the item all of these things can be turned off.
i also made it so the health bar in Blessing Of Knowledge changes color depending on how low health the entity is.